IMPALA-5172: fix incorrect cast in call to LZO decompress
For the call to lzo1x_decompress_safe, the output buffer
size that LZO expects is a lzo_uint*, which is a pointer
to a 64 bit unsigned integer. Our code currently uses
a pointer to uncompressed_len, which is a 32 bit integer.
This is incorrect and leads to a buffer overflow for
corrupted LZO files, because 4 extra bytes are included
in the output buffer size seen by LZO. This leads LZO to
consider the output buffer effectively unbounded. Any
corruption related to the size is not caught and can
read arbitrarily far beyond the end of the provided
buffer.

To fix this, the uncompressed_len is stored into a 64-bit
integer, which is passed into lzo1x_decompress_safe.
This also verifies that the expected uncompressed length
is greater than zero, as a negative uncompressed length
could lead to a buffer overrun.

Change-Id: Id46e7cb508c9e494718c329650361396698b1180