CDH-61317: don't return pointers to I/O buffers
The bug is that the scanner returned strings referencing I/O buffers
directly. The fix is to copy out uncompressed data into the same MemPool
as compressed data is decompressed into.

Testing:
Was able to reliably repro with the following sequence of commands:

  ./buildall.sh -asan -skiptests -noclean -ninja -notests
  start-impala-cluster.py --impalad_args=--disable_mem_pools=true
  impala-py.test tests/query_test/test_scanners.py -k lzo \
      --workload_exploration_strategy=functional-query:exhaustive

After this fix, it no longer reproduces.

We should get better coverage for this kind of bug once the I/O mgr is
switched over to the new buffer pool and can take advantage of ASAN
poisoning.

Change-Id: I838279bbe23ece6140b0e7ca74c563e8be6621ee