CDH-74236: Fix HdfsLzoTextScanner::FillByteBuffer with null MemPool
HdfsTextScanner::FillByteBuffer, which HdfsLzoTextScanner overrides,
takes a MemPool as a parameter. If a MemPool is provided,
FillByteBuffer should attach memory returned by previous calls to the
MemPool.

It is also valid to pass nullptr for the MemPool, in which case the
memory should just be freed. HdfsLzoTextScanner::FillByteBuffer did
not previously handle this case.

Testing:
- There is an existing test case that exercises this path in
  test_scan_ranges. It was previously not crashing due to bad luck.
  With a change that is about to go in to Impala (IMPALA-7272), it
  will start crashing without this fix.